fix(optimize-css): only remove `bx--` prefixed classes (#56)
Fixes #55

Currently, the `optimizeCss` plugin is too aggressive. It will remove non-Carbon prefixed selectors.

The fix is to first check that selectors actually include the Carbon `bx--` prefix. This is an even more conservative approach; however, the critical mandate is to avoid false positives.

With this change, unrelated selectors will not be removed:

* Scoped Svelte style
* Global Svelte style (e.g., :global())
* Global selector from CSS StyleSheet
* Global selector overriding a Carbon class